Add a campaign-csv-result service
This commit introduces a separate campaign csv result service. This
service's role is to define what are the csv columns we want to display,
given some campaign statistics defined in the CampaignIndividualResult
domain model object. It includes:
- a mapping between properties and csvcolumns
- how to react based on the availability of data
- the file name that we want to have
- the placeholder that we want to have.
